Eliminar el resource para procedures
